Decision Made
Noticed, using opentelementry was not ok with logging. The logs are not
parsed properly.

Resorted to using winston-loki

No exporter for metric, resorted to using prom-client.

Added services to the compose file and a prom config file.

Both was tested and data visualized in Grafana.

DCO Signed Off Samuel Ezeja <samuelemyrs@gmail.com>